Rename ConfigurationStepperButtons to configuration-stepper-buttons
