+++ packages/js/src/first-time-configuration/first-time-configuration-steps.js
@@ -8,7 +8,7 @@ import UnsavedChangesModal from "./tailwind-components/unsaved-changes-modal";
 import { addLinkToString } from "../helpers/stringHelpers.js";
 import SocialProfilesStep from "./tailwind-components/steps/social-profiles/social-profiles-step";
 import Stepper, { Step } from "./tailwind-components/Stepper";
-import { ContinueButton, EditButton, ConfigurationStepButtons } from "./tailwind-components/ConfigurationStepperButtons";
+import { ContinueButton, EditButton, ConfigurationStepButtons } from "./tailwind-components/configuration-stepper-buttons";
 import { getInitialActiveStepIndex } from "./stepper-helper";
 import IndexationStep from "./tailwind-components/steps/indexation/indexation-step";
 import SiteRepresentationStep from "./tailwind-components/steps/site-representation/site-representation-step";
