+++ packages/js/src/first-time-configuration/tailwind-components/configuration-stepper-buttons.js
